hopr_api

Module chain

On-chain operations-related API traits.

Structs

AccountEntry

Represents a node announcement entry on the block chain. This contains node’s public key and optional announcement information (multiaddress, block number).

AccountSelector

Selector for on-chain node accounts.

AcknowledgedTicket

Contains acknowledgment information and the respective ticket

Balance

Represents a non-negative balance of some Currency.

ChannelEntry

Overall description of a channel

ChannelSelector

Selector for channels.

DomainSeparators

Contains domain separator information.

HoprSphinxHeaderSpec

Current Sphinx header specification for the HOPR protocol.

Multiaddr

Representation of a Multiaddr.

SignificantChainEvent

Contains TX hash along with the Chain Event data. This could be used to pair up some events with Action. Each Action is typically concluded by a significant chain event.

Utc

The UTC time zone. This is the most efficient time zone when you don’t need the local time. It is also used as an offset (which is also a dummy type).

WinningProbability

Represents a ticket winning probability.

Enums

AnnouncementError

Error that can occur when making a node announcement.

ChainEventType

Enumeration of HOPR chain events.

ChannelDirection

Describes a direction of node’s own channel. The direction of a channel that is not own is undefined.

ChannelStatusDiscriminants

Traits

ChainEvents

Allows subscribing to on-chain events.

ChainKeyOperations

Operations for offchain keys.

ChainReadAccountOperations

Chain operations that read on-chain node accounts.

ChainReadChannelOperations

On-chain read operations regarding channels.

ChainValues

Retrieves various on-chain information.

ChainWriteAccountOperations

On-chain write operations regarding on-chain node accounts.

ChainWriteChannelOperations

On-chain write operations regarding channels.

ChainWriteTicketOperations

On-chain write operations with tickets.

Currency

Represents a general currency - like a token or a coin.

KeyIdMapper

Trait that defines 1:1 mapper between key identifiers and the actual public keys.

Type Aliases

ChainReceipt

Receipt of an on-chain operation.

ChannelId

Alias for the Hash representing a channel ID.

DateTime

HoprBalance

HoprSphinxSuite

Currently used public key cipher suite for Sphinx.