hopr_api::chain

Trait ChainReadAccountOperations

pub trait ChainReadAccountOperations {
    type Error: Error + Send + Sync + 'static;

    // Required methods
    fn node_balance<'life0, 'async_trait, C>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Balance<C>, Self::Error>> + Send + 'async_trait>>
       where C: 'async_trait + Currency,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn safe_balance<'life0, 'async_trait, C>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Balance<C>, Self::Error>> + Send + 'async_trait>>
       where C: 'async_trait + Currency,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn safe_allowance<'life0, 'async_trait, C>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Balance<C>, Self::Error>> + Send + 'async_trait>>
       where C: 'async_trait + Currency,
             Self: 'async_trait,
             'life0: 'async_trait;
    fn find_account_by_address<'life0, 'life1, 'async_trait>(
        &'life0 self,
        address: &'life1 Address,
    ) -> Pin<Box<dyn Future<Output = Result<Option<AccountEntry>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn find_account_by_packet_key<'life0, 'life1, 'async_trait>(
        &'life0 self,
        packet_key: &'life1 OffchainPublicKey,
    ) -> Pin<Box<dyn Future<Output = Result<Option<AccountEntry>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn check_node_safe_module_status<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn can_register_with_safe<'life0, 'life1, 'async_trait>(
        &'life0 self,
        safe_address: &'life1 Address,
    ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn stream_accounts<'a, 'async_trait>(
        &'a self,
        selector: AccountSelector,
    ) -> Pin<Box<dyn Future<Output = Result<BoxStream<'a, AccountEntry>, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait;
    fn count_accounts<'life0, 'async_trait>(
        &'life0 self,
        selector: AccountSelector,
    ) -> Pin<Box<dyn Future<Output = Result<usize, Self::Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Chain operations that read on-chain node accounts.

Required Associated Types

type Error: Error + Send + Sync + 'static

Required Methods

fn node_balance<'life0, 'async_trait, C>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Balance<C>, Self::Error>> + Send + 'async_trait>>
where C: 'async_trait + Currency, Self: 'async_trait, 'life0: 'async_trait,

Returns the native or token currency balance of the current node’s account.

fn safe_balance<'life0, 'async_trait, C>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Balance<C>, Self::Error>> + Send + 'async_trait>>
where C: 'async_trait + Currency, Self: 'async_trait, 'life0: 'async_trait,

Returns the native or token currency balance of the current node’s Safe.

fn safe_allowance<'life0, 'async_trait, C>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Balance<C>, Self::Error>> + Send + 'async_trait>>
where C: 'async_trait + Currency, Self: 'async_trait, 'life0: 'async_trait,

Returns the native or token currency Safe allowance.

fn find_account_by_address<'life0, 'life1, 'async_trait>( &'life0 self, address: &'life1 Address, ) -> Pin<Box<dyn Future<Output = Result<Option<AccountEntry>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns account entry given the on-chain address.

fn find_account_by_packet_key<'life0, 'life1, 'async_trait>( &'life0 self, packet_key: &'life1 OffchainPublicKey, ) -> Pin<Box<dyn Future<Output = Result<Option<AccountEntry>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Returns account entry given the off-chain packet_key.

fn check_node_safe_module_status<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Validates the node’s Safe setup.

fn can_register_with_safe<'life0, 'life1, 'async_trait>( &'life0 self, safe_address: &'life1 Address, ) -> Pin<Box<dyn Future<Output = Result<bool, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Checks if the given safe address can be registered with the current node.

fn stream_accounts<'a, 'async_trait>( &'a self, selector: AccountSelector, ) -> Pin<Box<dyn Future<Output = Result<BoxStream<'a, AccountEntry>, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait,

Returns on-chain node accounts with the given AccountSelector.

fn count_accounts<'life0, 'async_trait>( &'life0 self, selector: AccountSelector, ) -> Pin<Box<dyn Future<Output = Result<usize, Self::Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Counts the accounts with the given AccountSelector.

This is potentially done more effectively than counting more elements of the stream returned by ChainReadAccountOperations::stream_accounts.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.